feat(user-profile-page) Removed AccountInfoService
AB#23
